refactor: Active cube and arrows based on progress
