+++ apps/frontend/src/components/progressMap/progressMap.css
@@ -1,0 +1,68 @@
+.day-cell,
+.day-cell-side,
+.day-cell-bottom,
+.day-cell * {
+  overflow: visible;
+  transform-style: preserve-3d;
+}
+
+.day-cell {
+  position: relative;
+  transform-style: preserve-3d;
+  perspective: 1500px;
+  transform: translateZ(0);
+  transition: transform 0.25s ease;
+}
+
+.day-cell-side {
+  transform-style: preserve-3d;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%; /* cube thickness */
+  height: 100%;
+  background-color: black;
+  z-index: -10;
+  transform: scaleY(0);
+  transform-origin: left center;
+  transition: transform 0.25 ease;
+}
+
+.day-cell-bottom {
+  transform-style: preserve-3d;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: black;
+  z-index: -10;
+  transform: scaleY(0);
+  transform-origin: bottom center;
+  transition: transform 0.25 ease;
+}
+
+.day-cell:hover {
+  transform: translateY(-30px) translateX(30px);
+}
+
+.day-cell:hover .day-cell-side {
+  transform: scaleY(1) rotateY(112deg) skewY(20deg) translateY(1px);
+}
+
+.day-cell:hover .day-cell-bottom {
+  transform: scaleY(1) rotateX(111deg) skewX(19deg) translateY(4px)
+    translateZ(2px);
+}
+
+/* When Complete */
+.day-cell.day-complete {
+  transform: translateY(-30px) translateX(30px);
+}
+.day-cell.day-complete .day-cell-side {
+  transform: scaleY(1) rotateY(112deg) skewY(20deg) translateY(1px);
+}
+.day-cell.day-complete .day-cell-bottom {
+  transform: scaleY(1) rotateX(111deg) skewX(19deg) translateY(4px)
+    translateZ(2px);
+}
