+++ apps/frontend/src/pages/page.css
@@ -1,55 +1,3 @@
 .direction-reverse {
   direction: rtl;
 }
-
-.day-cell {
-  position: relative;
-  transform-style: preserve-3d;
-  perspective: 1500px;
-  transform: translateZ(0);
-  transition:
-    transform 0.25s ease,
-    box-shadow 0.25s ease;
-}
-
-.day-cell-side {
-  transform-style: preserve-3d;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%; /* cube thickness */
-  height: 100%;
-  background-color: black;
-  border: 5px solid gray;
-  z-index: -10;
-  transform: scaleX(0);
-  transform-origin: left center;
-  transition: transform 0.15 ease;
-}
-
-.day-cell-bottom {
-  transform-style: preserve-3d;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: black;
-  border: 5px solid gray;
-  z-index: -10;
-  transform: scaleX(0);
-  transform-origin: bottom center;
-  transition: transform 0.15 ease;
-}
-
-.day-cell:hover {
-  transform: translateY(-30px) translateX(30px);
-}
-
-.day-cell:hover .day-cell-side {
-  transform: scaleX(1) rotateY(115deg) rotateZ(15deg);
-}
-
-.day-cell:hover .day-cell-bottom {
-  transform: scaleX(1) rotateX(115deg) rotateZ(-25deg);
-}
